requirements_hotkey.txt
# macOS TTS Hotkey Application Dependencies

# Core TTS dependencies
torch>=1.9.0
torchaudio>=0.9.0
numpy>=1.21.0
kokoro-tts  # Main TTS library with KModel and KPipeline

# Audio processing and playback
pygame>=2.1.0
soundfile>=0.10.0

# System integration
pynput>=1.7.0
pyperclip>=1.8.0

# Text processing (if not already available)
phonemizer>=3.0.0
unidecode>=1.3.0

# Optional: for better audio handling
librosa>=0.8.0